refactor: highlight keys
-Refactor code that highlights calculator buttons/keys into a custom hook to be used universally to highlight keys. -Refactor calculaor button styling so that the handleBtnStyle component doesn't apply styles that don't need to be applied conditionally.
